Tarea 4: Segmentation faults evitados

+++ Shell_project.c
@@ -27,7 +27,7 @@ job *registroProcesos;
 
 
 void manejador(int senal) {
-    int status; // Estado del proceso, no vale para mucho
+    int status;                 // Estado del proceso, no vale para mucho
     enum status status_res;     // Resultado que dara el analyze_status
     int pid;                    // pid proceso a manejar
     int info;                   // Info necesaria
@@ -66,40 +66,52 @@ int selectComandoInterno(char **cmd) {
         }
         return 1;
     } else if (strcmp(cmd[0], "fg") == 0) {
-        int status;
-        int info;
-
         if (cmd[1] == NULL) cmd[1] = "1";
         sscanf(cmd[1], "%d", &i);
-        job *t = get_item_bypos(registroProcesos, i);
-        pid_t pg = t->pgid;
-        char copiaCMD[20];
-        strcpy(copiaCMD, t->command);
 
-        killpg(pg, SIGCONT);
+        if (empty_list(registroProcesos)) {
+            printf("No hay procesos suspendidos ni en segundo plano \n");
+        } else if (list_size(registroProcesos) > i) {
+            printf("No existe proceso en la posicion %d de la lista \n", i);
+        } else {
+            int status;
+            int info;
 
-        set_terminal(pg);
-        int pid_wait = waitpid(pg, &status, WUNTRACED);
-        delete_job(registroProcesos, t);
-        set_terminal(getpid());
+            job *t = get_item_bypos(registroProcesos, i);
+            pid_t pg = t->pgid;
+            char copiaCMD[20];
+            strcpy(copiaCMD, t->command);
 
-        enum status status_res = analyze_status(status, &info);
+            killpg(pg, SIGCONT);
 
-        if (strcmp(status_strings[status_res], "Suspended") == 0) {
-            job *suspendido = new_job(pid_wait, copiaCMD, STOPPED);
-            add_job(registroProcesos, suspendido);
+            set_terminal(pg);
+            int pid_wait = waitpid(pg, &status, WUNTRACED);
+            delete_job(registroProcesos, t);
+            set_terminal(getpid());
+
+            enum status status_res = analyze_status(status, &info);
+
+            if (strcmp(status_strings[status_res], "Suspended") == 0) {
+                job *suspendido = new_job(pid_wait, copiaCMD, STOPPED);
+                add_job(registroProcesos, suspendido);
+            }
         }
-
 
         return 1;
     } else if (strcmp(cmd[0], "bg") == 0) {
         if (cmd[1] == NULL) cmd[1] = "1";
         sscanf(cmd[1], "%d", &i);
-        job *t = get_item_bypos(registroProcesos, i);
-        pid_t pg = t->pgid;
-        t->state = BACKGROUND;
-        killpg(pg, SIGCONT);
 
+        if (empty_list(registroProcesos)) {
+            printf("No hay procesos suspendidos \n");
+        } else if (list_size(registroProcesos) > i) {
+            printf("No existe proceso en la posicion %d de la lista \n", i);
+        } else {
+            job *t = get_item_bypos(registroProcesos, i);
+            pid_t pg = t->pgid;
+            t->state = BACKGROUND;
+            killpg(pg, SIGCONT);
+        }
         return 1;
     }
